Changed interface tests from functional to unit
